Add TestAssertUnmovedNewActor to test module
