+++ rt/src/scheduler/tests.rs
@@ -1,19 +1,19 @@
 //! Tests for the local scheduler.
 
 use std::cell::RefCell;
+use std::future::pending;
 use std::future::Future;
-use std::future::{pending, Pending};
 use std::pin::Pin;
 use std::rc::Rc;
 use std::task::{self, Poll};
 
-use heph::actor::{self, ActorFuture, NewActor};
+use heph::actor::{self, ActorFuture};
 use heph::supervisor::NoSupervisor;
 
 use crate::process::{FutureProcess, Process, ProcessId};
 use crate::scheduler::{ProcessData, Scheduler};
 use crate::spawn::options::Priority;
-use crate::test::{self, assert_size, nop_task_waker, AssertUnmoved};
+use crate::test::{self, assert_size, nop_task_waker, AssertUnmoved, TestAssertUnmovedNewActor};
 use crate::ThreadLocal;
 
 #[test]
@@ -219,24 +219,6 @@ fn scheduler_run_order() {
     assert_eq!(*run_order.borrow(), vec![2_usize, 1, 0]);
 }
 
-struct TestAssertUnmovedNewActor;
-
-impl NewActor for TestAssertUnmovedNewActor {
-    type Message = ();
-    type Argument = ();
-    type Actor = AssertUnmoved<Pending<Result<(), !>>>;
-    type Error = !;
-    type RuntimeAccess = ThreadLocal;
-
-    fn new(
-        &mut self,
-        _: actor::Context<Self::Message, Self::RuntimeAccess>,
-        _: Self::Argument,
-    ) -> Result<Self::Actor, Self::Error> {
-        Ok(AssertUnmoved::new(pending()))
-    }
-}
-
 #[test]
 fn assert_actor_process_unmoved() {
     let mut scheduler = Scheduler::new();
@@ -244,7 +226,8 @@ fn assert_actor_process_unmoved() {
     let mut ctx = task::Context::from_waker(&waker);
 
     let rt = ThreadLocal::new(test::runtime());
-    let (process, _) = ActorFuture::new(NoSupervisor, TestAssertUnmovedNewActor, (), rt).unwrap();
+    let (process, _) =
+        ActorFuture::new(NoSupervisor, TestAssertUnmovedNewActor::new(), (), rt).unwrap();
     let pid = scheduler.add_new_process(Priority::NORMAL, process);
 
     // Run the process multiple times, ensure it's not moved in the process.
